Allow us to fudge a release when the local CPAN server has flaked.
Sometimes the CPAN::Mini::Inject::REST reports an error
despite successfully uploading and indexing a distribution.
this leaves us in an awkward state as we can't re-upload it
and we need this step to pass to get the rest of the release
tasks accomplished.

This allows us to silently pass by setting an environment variable.

ALREADY_UPLOADED_PRETEND_THIS_HAPPENED_ALREADY=1 dzil release